Floating-point registration marks.

+++ src/pockets/SpriteSheet.cpp
@@ -101,7 +101,7 @@ SpriteSheet::SpriteDataCollection SpriteSheet::defaultParseFunction( const JsonT
   {
     Rectf bounds( child["x1"].getValue<int>(), child["y1"].getValue<int>()
                  , child["x2"].getValue<int>(), child["y2"].getValue<int>() );
-    Vec2i registration_point( child["rx"].getValue<int>(), child["ry"].getValue<int>() );
+    Vec2i registration_point( child["rx"].getValue<float>(), child["ry"].getValue<float>() );
     string id = child["id"].getValue();
     SpriteData sprite( bounds.getSize()
                      , Rectf( bounds.getUpperLeft() / bitmap_size
